Removed some buf_write_t for superblocks initialization
Added specific application of a virtual buf_write_t operation in the
buf_lock_t constructor.  This is to prevent a crash caused by the
breaking of the invariant that every newly created block has a
buf_write_t operation.  (This has something to do with the recency
being left with the value repli_timestamp_t::invalid -- no index write
to disk?  Maybe.)