TableViewer WIP
Moved main point of development to tables page, instead of data
explorer.

Made TableRowSource to be the row source for tables.

Made scrolling use preloading.

Made TableViewer discard DOM elements that aren't necessary for
rendering in the current scroll position.  When scrolling up, we jump
upwards very quickly, and when scrolling down, we get jumpy
adjustments of the scrollbar size.  We need to add proper
margin/padding (to something) to get the scroll thumb to behave
somewhat smoothly.